Title: 삼 오 무, Time: 3034ms, Memory: 8MB, Status: Time Limit Exceed - CodeTree

+++ 240529/삼 오 무/three-five-moo.kt
@@ -4,19 +4,19 @@ import java.io.*
 class Main {
     private val br = BufferedReader(InputStreamReader(System.`in`))
 
-    private var n = 0L
+    private var n = 0
 
     fun solve() {
-        n = br.readLine().toLong()
+        n = br.readLine().toInt()
 
         println(parametricSearch())
     }
 
-    private fun parametricSearch() : Long {
-        var start = 1L
-        var end = (Int.MAX_VALUE).toLong()
+    private fun parametricSearch() : Int {
+        var start = 1
+        var end = Int.MAX_VALUE
 
-        var ret = Long.MAX_VALUE
+        var ret = Int.MAX_VALUE
 
         while(start <= end) {
             val mid = (start + end) / 2
@@ -32,7 +32,7 @@ class Main {
         return ret
     }
 
-    private fun cnt(num : Long) : Long {
+    private fun cnt(num : Int) : Int {
         return num - num / 3 - num / 5 + num / 15
     }
 }
